algo require holidays and manage holidays, not sending email yet

+++ back/routes/controllers/demandes.js
@@ -4,11 +4,13 @@ const { response } = require('../../app');
 const router = express.Router();
 const { Demandes } = require("../../models");
 const { Holidays } = require("../../models");
+const { Types } = require("../../models/Types.js");
+const { Users } = require("../../models");
 // to create demande
 exports.createDemande = async (req, res) => {
     try{
         const demande = req.body;
-        demande.status = 1;
+        demande.idStatus = 1;
         const idUser = req.body.idUser;
         
         // if this user's available holidays are not 0, create demande
@@ -18,8 +20,6 @@ exports.createDemande = async (req, res) => {
         
         const daysDemande = Math.ceil((date2.getTime() - date1.getTime())/ (1000 * 3600 * 24)+1);
        // res.json(daysDemande)
-
-        
         if(demande.idType ==1 && holiday.holidaysAvailable >= daysDemande){
             await Demandes.create(demande).then(createdDemande=>{
                 res.json(createdDemande); // return created demande
@@ -28,9 +28,9 @@ exports.createDemande = async (req, res) => {
             res.json("you dont have available holidays");
         } else if(demande.idType != 1){
             // create demande
-                await Demandes.create(demande).then(createdDemande=>{
-                    res.json(createdDemande); // return created demande 
-                }); 
+            await Demandes.create(demande).then(createdDemande=>{
+                res.json(createdDemande); // return created demande 
+            }); 
         }
     } catch (error) {
         res.send(error);
@@ -64,7 +64,7 @@ exports.getDemandeByIdUser = async (req, res) => {
             where: {
                 idUser: [id]
             }
-        });;
+        });
         res.json(demandes); 
     }catch (error) {
         res.send(error);
@@ -77,14 +77,20 @@ exports.deleteDemandeById = async (req, res) => {
         const demande = await Demandes.findByPk(id);
         // if this user has holidays, delete his holidays
         if(demande){ 
-            await Demandes.destroy({
-                where: {
-                    id: [id]
-                }
-            });
+            if(demande.idStatus !=2 ){
+                await Demandes.destroy({
+                    where: {
+                        id: [id]
+                    }
+                });
+                const demandes = await Demandes.findAll();
+                res.json(demandes);
+            } else {
+                res.json("demande accepted, can not be deleted");
+            }    
+        }else {
+            res.json("demande doesn't exist");
         }
-        // delete this user
-        res.json("demande deleted");
     }catch (error) {
         res.send(error);
     }
@@ -93,9 +99,14 @@ exports.deleteDemandeById = async (req, res) => {
 exports.updateDemande = async (req, res) => {
     try {
         const id = req.body.id;
-        const status = req.body.status;
+        const demandeOriginal = await Demandes.findByPk(id);
+        const objDemande = req.body; 
+        const iduser = demandeOriginal.idUser;
+        const idtype = demandeOriginal.idType;
+        const status = req.body.idStatus;
         const description = req.body.description;
-        const objDemande = req.body;
+        
+        
         delete objDemande.id;
         if(status == 3){
             if(!description){
@@ -104,17 +115,42 @@ exports.updateDemande = async (req, res) => {
                 await Demandes.update(objDemande, {
                     where : {id: [id]}
                 });
-                const demande = await Demandes.findByPk(id);
-                res.json(demande);
+                const demandes = await Demandes.findByPk(id);
+                res.json(demandes);
             }
-        } else {
-            await Demandes.update(objDemande, {
+        } else if(status == 2 && idtype == 1 ){ // if congé payé normale accepted
+            await Demandes.update(objDemande, { // update demande status
                 where : {id: [id]}
+            });
+            // update holiday
+            const holiday = await Holidays.findOne({
+                where: {idUser: iduser}});
+            const date1 = new Date(demandeOriginal.startingDate);
+            const date2 = new Date(demandeOriginal.endingDate);     
+            const daysDemande = Math.ceil((date2.getTime() - date1.getTime())/ (1000 * 3600 * 24)+1);
+            const holidaysAvb = holiday.holidaysAvailable-daysDemande;
+            const holidaysTaken = holiday.holidaysTaken + daysDemande;
+            const holidayUpdate = {
+                "holidaysAvailable": holidaysAvb, 
+                "holidaysTaken": holidaysTaken
+                };
+            await Holidays.update(holidayUpdate,{    // update user
+                where: { 
+                    idUser : iduser
+                }, 
             });
             const demande = await Demandes.findByPk(id);
                 res.json(demande);
+        } else if (status == 2 && idtype != 1 ) { // if other type(maladie...) congé payé accepted
+            await Demandes.update(objDemande, { // update demande status
+                where : {id: [id]}
+            });
+            const demande = await Demandes.findByPk(id);
+            res.json(demande);
+        } else if (status ==1){
+            res.json("status did not change");
         }
     }catch (error) {
         res.send(error);
     }
-}
+};
